+++ src/BankingSystem.Api/Application/Factories/Implementations/CDT.cs
@@ -1,15 +1,15 @@
 ﻿namespace BankingSystem.Api.Application.Factories.Implementations;
 public class CDT : IAccount
 {
-    private int _NUMBER;
-    public CDT(int number, double balance)
+    private long _NUMBER;
+    public CDT(long number, double balance)
     {
         _NUMBER = number;
         Balance = balance;
     }
 
     public double Balance { get; private set; }
-    public int Number => _NUMBER;
-    public int GetNumber() => Number;
+    public long Number => _NUMBER;
+    public long GetNumber() => Number;
     public double GetBalance() => Balance;
 }
