+++ src/BankingSystem.Api/Application/Factories/Implementations/CDTFactory.cs
@@ -4,7 +4,7 @@ public class CDTFactory : IBankingProductFactory
 {
     public IAccount CreateAccount(IAccountNumberGeneratorService accountNumberGeneratorService, double initialBalance = 0)
     {
-        int number = accountNumberGeneratorService.GenerateNumber();
+        long number = accountNumberGeneratorService.GenerateNumber();
 
         ValidateInitialBalance(initialBalance);
