+++ src/BankingSystem.Api/Application/Factories/Implementations/CheckingAccount.cs
@@ -1,16 +1,16 @@
 ﻿namespace BankingSystem.Api.Application.Factories.Implementations;
 public class CheckingAccount : IAccount
 {
-    private int _NUMBER;
+    private long _NUMBER;
 
-    public CheckingAccount(int number, double initialBalance)
+    public CheckingAccount(long number, double initialBalance)
     {
         _NUMBER = number;
         Balance = initialBalance;
     }
 
     public double Balance { get; private set; }
-    public int Number => _NUMBER;
-    public int GetNumber() => Number;
+    public long Number => _NUMBER;
+    public long GetNumber() => Number;
     public double GetBalance() => Balance;
 }
