+++ src/BankingSystem.Api/Application/Factories/Implementations/CheckingAccountFactory.cs
@@ -3,7 +3,7 @@ public class CheckingAccountFactory : IBankingProductFactory
 {
     public IAccount CreateAccount(IAccountNumberGeneratorService accountNumberGeneratorService, double initialBalance = 0)
     {
-        int number = accountNumberGeneratorService.GenerateNumber();
+        long number = accountNumberGeneratorService.GenerateNumber();
 
         var account = new CheckingAccount(number, initialBalance);
         return account;
