+++ src/BankingSystem.Api/Application/Factories/Implementations/SavingsAccount.cs
@@ -1,16 +1,16 @@
 ﻿namespace BankingSystem.Api.Application.Factories.Implementations;
 public class SavingsAccount : IAccount
 {
-    private int _NUMBER;
+    private long _NUMBER;
 
-    public SavingsAccount(int number, double initialBalance)
+    public SavingsAccount(long number, double initialBalance)
     {
         _NUMBER = number;
         Balance = initialBalance;
     }
 
-    public int Number => _NUMBER;
+    public long Number => _NUMBER;
     public double Balance { get; private set; }
-    public int GetNumber() => Number;
+    public long GetNumber() => Number;
     public double GetBalance() => Balance;
 }
