+++ src/BankingSystem.Api/Application/Factories/Implementations/SavingsAccountFactory.cs
@@ -3,7 +3,7 @@ public class SavingsAccountFactory : IBankingProductFactory
 {
     public IAccount CreateAccount(IAccountNumberGeneratorService accountNumberGeneratorService, double initialBalance = 0)
     {
-        int number = accountNumberGeneratorService.GenerateNumber();
+        long number = accountNumberGeneratorService.GenerateNumber();
 
         var account = new SavingsAccount(number, initialBalance);
         return account;
