+++ src/BankingSystem.Infrastructure/Services/IAccountNumberGeneratorService.cs
@@ -1,5 +1,5 @@
 ﻿namespace BankingSystem.Infrastructure.Services;
 public interface IAccountNumberGeneratorService
 {
-    int GenerateNumber();
+    long GenerateNumber();
 }
